移除 grey_theory.py 並整合到 __init__.py 中
